// frontend/src/components/layout/types.ts
import { type LinkProps } from '@tanstack/react-router'

type User = {
  name: string
  email: string
  avatar: string
}

type Team = {
  name: string
  logo: React.ElementType
  plan: string
}

type BaseNavItem = {
  title: string
  badge?: string
  icon?: string
}

type NavLink = BaseNavItem & {
  url: LinkProps['to'] | (string & {})
  is_frame?: boolean
  items?: never
}

type NavCollapsible = BaseNavItem & {
  items: NavItem[]
  url?: never
}

type NavItem = NavCollapsible | NavLink

type NavGroup = {
  title: string
  icon?: string
  items?: NavItem[]
  url?: LinkProps['to'] | (string & {})
  badge?: string
  is_frame?: boolean
}

type SidebarData = {
  user: User
  teams: Team[]
  navGroups: NavGroup[]
}

export type { SidebarData, NavGroup, NavItem, NavCollapsible, NavLink }